+++ pkg/grpc/service/study-management_endpoints.go
@@ -3,12 +3,16 @@ package service
 import (
 	"context"
 	"fmt"
+	"reflect"
+	"time"
 
 	"github.com/influenzanet/go-utils/pkg/api_types"
 	"github.com/influenzanet/go-utils/pkg/constants"
 	"github.com/influenzanet/go-utils/pkg/token_checks"
 	loggingAPI "github.com/influenzanet/logging-service/pkg/api"
 	"github.com/influenzanet/study-service/pkg/api"
+	"github.com/influenzanet/study-service/pkg/dbs/studydb"
+	"github.com/influenzanet/study-service/pkg/studyengine"
 	"github.com/influenzanet/study-service/pkg/types"
 	"github.com/influenzanet/study-service/pkg/utils"
 	"google.golang.org/grpc/codes"
@@ -125,7 +129,7 @@ func (s *studyServiceServer) SaveSurveyToStudy(ctx context.Context, req *api.Add
 
 	newSurvey := types.SurveyFromAPI(req.Survey)
 	if newSurvey.Current.VersionID == "" {
-		newSurvey.Current.VersionID = utils.GenerateSurveyVersionID()
+		newSurvey.Current.VersionID = utils.GenerateSurveyVersionID(newSurvey.History)
 	}
 	createdSurvey, err := s.studyDBservice.SaveSurvey(req.Token.InstanceId, req.StudyKey, newSurvey)
 	if err != nil {
@@ -381,6 +385,83 @@ func (s *studyServiceServer) SaveStudyProps(ctx context.Context, req *api.StudyP
 	}
 	s.SaveLogEvent(req.Token.InstanceId, req.Token.Id, loggingAPI.LogEventType_LOG, constants.LOG_EVENT_STUDY_UPDATE, req.StudyKey)
 	return uStudy.ToAPI(), nil
+}
+
+func (s *studyServiceServer) RunRules(ctx context.Context, req *api.StudyRulesReq) (*api.RuleRunSummary, error) {
+	if req == nil || token_checks.IsTokenEmpty(req.Token) || req.StudyKey == "" {
+		return nil, status.Error(codes.InvalidArgument, "missing argument")
+	}
+
+	if !token_checks.CheckRoleInToken(req.Token, constants.USER_ROLE_ADMIN) {
+		err := s.HasRoleInStudy(req.Token.InstanceId, req.StudyKey, req.Token.Id,
+			[]string{types.STUDY_ROLE_MAINTAINER, types.STUDY_ROLE_OWNER},
+		)
+		if err != nil {
+			s.SaveLogEvent(req.Token.InstanceId, req.Token.Id, loggingAPI.LogEventType_SECURITY, constants.LOG_EVENT_RUN_CUSTOM_RULES, fmt.Sprintf("permission denied for running custom rules in study %s  ", req.StudyKey))
+			return nil, status.Error(codes.Internal, err.Error())
+		}
+	}
+
+	start := time.Now().Unix()
+
+	type Counters struct {
+		Participants                  int32
+		ParticipantStateChangePerRule []int32
+	}
+	counters := &Counters{
+		Participants:                  0,
+		ParticipantStateChangePerRule: make([]int32, len(req.Rules)),
+	}
+
+	s.studyDBservice.FindAndExecuteOnParticipantsStates(
+		req.Token.InstanceId,
+		req.StudyKey,
+		"",
+		func(dbService *studydb.StudyDBService, p types.ParticipantState, instanceID, studyKey string, args ...interface{}) error {
+			counters.Participants += 1
+
+			currentState := p
+			for index, rule := range req.Rules {
+				exp := types.ExpressionFromAPI(rule)
+				if exp == nil {
+					continue
+				}
+
+				event := types.StudyEvent{
+					InstanceID: instanceID,
+					StudyKey:   studyKey,
+				}
+				newState, err := studyengine.ActionEval(*exp, currentState, event, s.studyDBservice)
+				if err != nil {
+					return err
+				}
+
+				if !reflect.DeepEqual(newState, currentState) {
+					counters.ParticipantStateChangePerRule[index] += 1
+				}
+				currentState = newState
+			}
+
+			if !reflect.DeepEqual(p, currentState) {
+				// save state back to DB
+				_, err := s.studyDBservice.SaveParticipantState(instanceID, studyKey, currentState)
+				if err != nil {
+					return status.Error(codes.Internal, err.Error())
+				}
+
+			}
+
+			return nil
+		},
+	)
+
+	s.SaveLogEvent(req.Token.InstanceId, req.Token.Id, loggingAPI.LogEventType_LOG, constants.LOG_EVENT_RUN_CUSTOM_RULES, fmt.Sprintf("rules run for study %s: %v", req.StudyKey, req.Rules))
+	resp := api.RuleRunSummary{
+		ParticipantCount:              counters.Participants,
+		ParticipantStateChangePerRule: counters.ParticipantStateChangePerRule,
+		Duration:                      time.Now().Unix() - start,
+	}
+	return &resp, nil
 }
 
 func (s *studyServiceServer) DeleteStudy(ctx context.Context, req *api.StudyReferenceReq) (*api.ServiceStatus, error) {