+++ pkg/grpc/service/studyflow_endpoints.go
@@ -49,7 +49,9 @@ func (s *studyServiceServer) EnterStudy(ctx context.Context, req *api.EnterStudy
 
 	// perform study rules/actions
 	currentEvent := types.StudyEvent{
-		Type: "ENTER",
+		Type:       "ENTER",
+		InstanceID: req.Token.InstanceId,
+		StudyKey:   req.StudyKey,
 	}
 	pState, err = s.getAndPerformStudyRules(req.Token.InstanceId, req.StudyKey, pState, currentEvent)
 	if err != nil {
@@ -229,8 +231,10 @@ func (s *studyServiceServer) PostponeSurvey(ctx context.Context, req *api.Postpo
 					}
 					// perform study rules/actions
 					currentEvent := types.StudyEvent{
-						Type:     "SUBMIT",
-						Response: emptyResponse,
+						Type:       "SUBMIT",
+						Response:   emptyResponse,
+						InstanceID: req.Token.InstanceId,
+						StudyKey:   req.StudyKey,
 					}
 					pState, err = s.getAndPerformStudyRules(req.Token.InstanceId, req.StudyKey, pState, currentEvent)
 					if err != nil {
@@ -304,8 +308,10 @@ func (s *studyServiceServer) SubmitStatusReport(ctx context.Context, req *api.St
 
 		// perform study rules/actions
 		currentEvent := types.StudyEvent{
-			Type:     "SUBMIT",
-			Response: response,
+			Type:       "SUBMIT",
+			Response:   response,
+			InstanceID: req.Token.InstanceId,
+			StudyKey:   study.Key,
 		}
 		pState, err = s.getAndPerformStudyRules(req.Token.InstanceId, study.Key, pState, currentEvent)
 		if err != nil {
@@ -361,8 +367,10 @@ func (s *studyServiceServer) SubmitResponse(ctx context.Context, req *api.Submit
 
 	// perform study rules/actions
 	currentEvent := types.StudyEvent{
-		Type:     "SUBMIT",
-		Response: response,
+		Type:       "SUBMIT",
+		Response:   response,
+		InstanceID: req.Token.InstanceId,
+		StudyKey:   req.StudyKey,
 	}
 	pState, err = s.getAndPerformStudyRules(req.Token.InstanceId, req.StudyKey, pState, currentEvent)
 	if err != nil {
@@ -418,7 +426,9 @@ func (s *studyServiceServer) LeaveStudy(ctx context.Context, req *api.LeaveStudy
 	}
 	// perform study rules/actions
 	currentEvent := types.StudyEvent{
-		Type: "LEAVE",
+		Type:       "LEAVE",
+		InstanceID: req.Token.InstanceId,
+		StudyKey:   req.StudyKey,
 	}
 	pState, err = s.getAndPerformStudyRules(req.Token.InstanceId, req.StudyKey, pState, currentEvent)
 	if err != nil {