+++ pkg/types/study-event.go
@@ -1,6 +1,8 @@
 package types
 
 type StudyEvent struct {
-	Type     string         // what kind of event (TIMER, SUBMISSION, ENTER etc.)
-	Response SurveyResponse // if something is submitted during the event is added here
+	InstanceID string
+	StudyKey   string
+	Type       string         // what kind of event (TIMER, SUBMISSION, ENTER etc.)
+	Response   SurveyResponse // if something is submitted during the event is added here
 }