+++ pkg/utils/utils_test.go
@@ -2,23 +2,23 @@ package utils
 
 import (
 	"testing"
-	"time"
+
+	"github.com/influenzanet/study-service/pkg/types"
 )
 
 func TestGenerateSurveyVersionID(t *testing.T) {
 	t.Run("test id generation for uniqueness", func(t *testing.T) {
-		ids := []string{}
+		oldVersions := []types.SurveyVersion{}
 
 		for i := 0; i < 100; i++ {
-			id := GenerateSurveyVersionID()
-			ids = append(ids, id)
-			time.Sleep(time.Millisecond)
+			id := GenerateSurveyVersionID(oldVersions)
+			oldVersions = append(oldVersions, types.SurveyVersion{VersionID: id})
 		}
 
-		for i, id_1 := range ids {
-			for j, id_2 := range ids {
-				if i != j && id_1 == id_2 {
-					t.Errorf("duplicate key present: i: %d - %s j: %d - %s ", i, id_1, j, id_2)
+		for i, id_1 := range oldVersions {
+			for j, id_2 := range oldVersions {
+				if i != j && id_1.VersionID == id_2.VersionID {
+					t.Errorf("duplicate key present: i: %d - %s j: %d - %s ", i, id_1.VersionID, j, id_2.VersionID)
 				}
 			}
 		}